+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [Change resolutions] - 2025-03-01
 
 * Added the ability to change resolutions.
+* Added the ability to change brightness.
 
 ## [Initial Version] - 2024-02-28